abstract helpers properly
