+++ validator/slashing-protection/local/standard-protection-format/BUILD.bazel
@@ -42,6 +42,7 @@ go_test(
         "//shared/testutil/require:go_default_library",
         "//validator/db/kv:go_default_library",
         "//validator/db/testing:go_default_library",
+        "//validator/testing:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
