+++ validator/slashing-protection/local/standard-protection-format/import_test.go
@@ -10,13 +10,8 @@ import (
 	"testing"
 
 	"github.com/prysmaticlabs/prysm/shared/bls"
-	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/hashutil"
-	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/prysmaticlabs/prysm/shared/rand"
-	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
-	"github.com/prysmaticlabs/prysm/validator/db/kv"
 	dbtest "github.com/prysmaticlabs/prysm/validator/db/testing"
 	logTest "github.com/sirupsen/logrus/hooks/test"
 )
@@ -43,103 +38,6 @@ func TestStore_ImportInterchangeData_NilData_FailsSilently(t *testing.T) {
 	err = ImportStandardProtectionJSON(ctx, validatorDB, buf)
 	require.NoError(t, err)
 	require.LogsContain(t, hook, "No slashing protection data to import")
-}
-
-func TestStore_ImportInterchangeData_BadFormat_PreventsDBWrites(t *testing.T) {
-	ctx := context.Background()
-	numValidators := 10
-	publicKeys := createRandomPubKeys(t, numValidators)
-	validatorDB := dbtest.SetupDB(t, publicKeys)
-
-	// First we setup some mock attesting and proposal histories and create a mock
-	// standard slashing protection format JSON struct.
-	attestingHistory, proposalHistory := mockAttestingAndProposalHistories(t, numValidators)
-	standardProtectionFormat := mockSlashingProtectionJSON(t, publicKeys, attestingHistory, proposalHistory)
-
-	// We replace a slot of one of the blocks with junk data.
-	standardProtectionFormat.Data[0].SignedBlocks[0].Slot = "BadSlot"
-
-	// We encode the standard slashing protection struct into a JSON format.
-	blob, err := json.Marshal(standardProtectionFormat)
-	require.NoError(t, err)
-	buf := bytes.NewBuffer(blob)
-
-	// Next, we attempt to import it into our validator database and check that
-	// we obtain an error during the import process.
-	err = ImportStandardProtectionJSON(ctx, validatorDB, buf)
-	assert.NotNil(t, err)
-
-	// Next, we attempt to retrieve the attesting and proposals histories from our database and
-	// verify nothing was saved to the DB. If there is an error in the import process, we need to make
-	// sure writing is an atomic operation: either the import succeeds and saves the slashing protection
-	// data to our DB, or it does not.
-	receivedAttestingHistory, err := validatorDB.AttestationHistoryForPubKeysV2(ctx, publicKeys)
-	require.NoError(t, err)
-	for i := 0; i < len(publicKeys); i++ {
-		defaultAttestingHistory := kv.NewAttestationHistoryArray(0)
-		require.DeepEqual(
-			t,
-			defaultAttestingHistory,
-			receivedAttestingHistory[publicKeys[i]],
-			"Imported attestation protection history is different than the empty default",
-		)
-		proposals := proposalHistory[i].Proposals
-		for _, proposal := range proposals {
-			receivedProposalSigningRoot, _, err := validatorDB.ProposalHistoryForSlot(ctx, publicKeys[i], proposal.Slot)
-			require.NoError(t, err)
-			require.DeepEqual(
-				t,
-				params.BeaconConfig().ZeroHash,
-				receivedProposalSigningRoot,
-				"Imported proposal signing root is different than the empty default",
-			)
-		}
-	}
-}
-
-func TestStore_ImportInterchangeData_OK(t *testing.T) {
-	ctx := context.Background()
-	numValidators := 10
-	publicKeys := createRandomPubKeys(t, numValidators)
-	validatorDB := dbtest.SetupDB(t, publicKeys)
-
-	// First we setup some mock attesting and proposal histories and create a mock
-	// standard slashing protection format JSON struct.
-	attestingHistory, proposalHistory := mockAttestingAndProposalHistories(t, numValidators)
-	standardProtectionFormat := mockSlashingProtectionJSON(t, publicKeys, attestingHistory, proposalHistory)
-
-	// We encode the standard slashing protection struct into a JSON format.
-	blob, err := json.Marshal(standardProtectionFormat)
-	require.NoError(t, err)
-	buf := bytes.NewBuffer(blob)
-
-	// Next, we attempt to import it into our validator database.
-	err = ImportStandardProtectionJSON(ctx, validatorDB, buf)
-	require.NoError(t, err)
-
-	// Next, we attempt to retrieve the attesting and proposals histories from our database and
-	// verify those indeed match the originally generated mock histories.
-	receivedAttestingHistory, err := validatorDB.AttestationHistoryForPubKeysV2(ctx, publicKeys)
-	require.NoError(t, err)
-	for i := 0; i < len(publicKeys); i++ {
-		require.DeepEqual(
-			t,
-			attestingHistory[i],
-			receivedAttestingHistory[publicKeys[i]],
-			"We should have stored any attesting history",
-		)
-		proposals := proposalHistory[i].Proposals
-		for _, proposal := range proposals {
-			receivedProposalSigningRoot, _, err := validatorDB.ProposalHistoryForSlot(ctx, publicKeys[i], proposal.Slot)
-			require.NoError(t, err)
-			require.DeepEqual(
-				t,
-				receivedProposalSigningRoot[:],
-				proposal.SigningRoot,
-				"Imported proposals are different then the generated ones",
-			)
-		}
-	}
 }
 
 func Test_validateMetadata(t *testing.T) {
@@ -771,84 +669,6 @@ func Test_saveLowestSourceTargetToDBt_Ok(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, true, e)
 	require.Equal(t, uint64(6), got)
-}
-
-func mockSlashingProtectionJSON(
-	t *testing.T,
-	publicKeys [][48]byte,
-	attestingHistories []kv.EncHistoryData,
-	proposalHistories []kv.ProposalHistoryForPubkey,
-) *EIPSlashingProtectionFormat {
-	standardProtectionFormat := &EIPSlashingProtectionFormat{}
-	standardProtectionFormat.Metadata.GenesisValidatorsRoot = fmt.Sprintf("%#x", bytesutil.PadTo([]byte{32}, 32))
-	standardProtectionFormat.Metadata.InterchangeFormatVersion = INTERCHANGE_FORMAT_VERSION
-	ctx := context.Background()
-	for i := 0; i < len(publicKeys); i++ {
-		data := &ProtectionData{
-			Pubkey: fmt.Sprintf("%#x", publicKeys[i]),
-		}
-		highestEpochWritten, err := attestingHistories[i].GetLatestEpochWritten(ctx)
-		require.NoError(t, err)
-		for target := uint64(0); target <= highestEpochWritten; target++ {
-			hd, err := attestingHistories[i].GetTargetData(ctx, target)
-			require.NoError(t, err)
-			data.SignedAttestations = append(data.SignedAttestations, &SignedAttestation{
-				TargetEpoch: fmt.Sprintf("%d", target),
-				SourceEpoch: fmt.Sprintf("%d", hd.Source),
-				SigningRoot: fmt.Sprintf("%#x", hd.SigningRoot),
-			})
-		}
-		for target := uint64(0); target < highestEpochWritten; target++ {
-			proposal := proposalHistories[i].Proposals[target]
-			block := &SignedBlock{
-				Slot:        fmt.Sprintf("%d", proposal.Slot),
-				SigningRoot: fmt.Sprintf("%#x", proposal.SigningRoot),
-			}
-			data.SignedBlocks = append(data.SignedBlocks, block)
-
-		}
-		standardProtectionFormat.Data = append(standardProtectionFormat.Data, data)
-	}
-	return standardProtectionFormat
-}
-
-func mockAttestingAndProposalHistories(t *testing.T, numValidators int) ([]kv.EncHistoryData, []kv.ProposalHistoryForPubkey) {
-	// deduplicate and transform them into our internal format.
-	attData := make([]kv.EncHistoryData, numValidators)
-	proposalData := make([]kv.ProposalHistoryForPubkey, numValidators)
-	gen := rand.NewGenerator()
-	ctx := context.Background()
-	for v := 0; v < numValidators; v++ {
-		var err error
-		latestTarget := gen.Intn(int(params.BeaconConfig().WeakSubjectivityPeriod) / 1000)
-		hd := kv.NewAttestationHistoryArray(uint64(latestTarget))
-		proposals := make([]kv.Proposal, 0)
-		for i := 1; i < latestTarget; i++ {
-			signingRoot := [32]byte{}
-			signingRootStr := fmt.Sprintf("%d", i)
-			copy(signingRoot[:], signingRootStr)
-			historyData := &kv.HistoryData{
-				Source:      uint64(gen.Intn(100000)),
-				SigningRoot: signingRoot[:],
-			}
-			hd, err = hd.SetTargetData(ctx, uint64(i), historyData)
-			require.NoError(t, err)
-		}
-		for i := 1; i <= latestTarget; i++ {
-			signingRoot := [32]byte{}
-			signingRootStr := fmt.Sprintf("%d", i)
-			copy(signingRoot[:], signingRootStr)
-			proposals = append(proposals, kv.Proposal{
-				Slot:        uint64(i),
-				SigningRoot: signingRoot[:],
-			})
-		}
-		proposalData[v] = kv.ProposalHistoryForPubkey{Proposals: proposals}
-		hd, err = hd.SetLatestEpochWritten(ctx, uint64(latestTarget))
-		require.NoError(t, err)
-		attData[v] = hd
-	}
-	return attData, proposalData
 }
 
 func createRandomPubKeys(t *testing.T, numValidators int) [][48]byte {
